+++ src/game/Object/Player.cpp
@@ -6359,100 +6359,6 @@ void Player::UpdateDefense()
 }
 
 /**
- * @brief Applies or removes a base modifier affecting derived combat values.
- *
- * @param modGroup The modifier group to update.
- * @param modType The modifier type to apply.
- * @param amount The modifier amount.
- * @param apply True to apply the modifier; false to remove it.
- */
-void Player::HandleBaseModValue(BaseModGroup modGroup, BaseModType modType, float amount, bool apply)
-{
-    if (modGroup >= BASEMOD_END || modType >= MOD_END)
-    {
-        sLog.outError("ERROR in HandleBaseModValue(): nonexistent BaseModGroup of wrong BaseModType!");
-        return;
-    }
-
-    float val = 1.0f;
-
-    switch (modType)
-    {
-        case FLAT_MOD:
-            m_auraBaseMod[modGroup][modType] += apply ? amount : -amount;
-            break;
-        case PCT_MOD:
-            if (amount <= -100.0f)
-            {
-                amount = -200.0f;
-            }
-
-            val = (100.0f + amount) / 100.0f;
-            m_auraBaseMod[modGroup][modType] *= apply ? val : (1.0f / val);
-            break;
-    }
-
-    if (!CanModifyStats())
-    {
-        return;
-    }
-
-    switch (modGroup)
-    {
-        case CRIT_PERCENTAGE:              UpdateCritPercentage(BASE_ATTACK);                          break;
-        case RANGED_CRIT_PERCENTAGE:       UpdateCritPercentage(RANGED_ATTACK);                        break;
-        case OFFHAND_CRIT_PERCENTAGE:      UpdateCritPercentage(OFF_ATTACK);                           break;
-        case SHIELD_BLOCK_VALUE:           UpdateShieldBlockValue();                                   break;
-        default: break;
-    }
-}
-
-/**
- * @brief Gets a stored base modifier value.
- *
- * @param modGroup The modifier group to query.
- * @param modType The modifier type to query.
- * @return The stored modifier value.
- */
-float Player::GetBaseModValue(BaseModGroup modGroup, BaseModType modType) const
-{
-    if (modGroup >= BASEMOD_END || modType > MOD_END)
-    {
-        sLog.outError("trial to access nonexistent BaseModGroup or wrong BaseModType!");
-        return 0.0f;
-    }
-
-    if (modType == PCT_MOD && m_auraBaseMod[modGroup][PCT_MOD] <= 0.0f)
-    {
-        return 0.0f;
-    }
-
-    return m_auraBaseMod[modGroup][modType];
-}
-
-/**
- * @brief Gets the combined flat and percentage base modifier value for a group.
- *
- * @param modGroup The modifier group to query.
- * @return The total effective modifier value.
- */
-float Player::GetTotalBaseModValue(BaseModGroup modGroup) const
-{
-    if (modGroup >= BASEMOD_END)
-    {
-        sLog.outError("wrong BaseModGroup in GetTotalBaseModValue()!");
-        return 0.0f;
-    }
-
-    if (m_auraBaseMod[modGroup][PCT_MOD] <= 0.0f)
-    {
-        return 0.0f;
-    }
-
-    return m_auraBaseMod[modGroup][FLAT_MOD] * m_auraBaseMod[modGroup][PCT_MOD];
-}
-
-/**
  * @brief Computes the player's current shield block value.
  *
  * @return The effective shield block amount.
@@ -6464,126 +6370,6 @@ uint32 Player::GetShieldBlockValue() const
     value = (value < 0) ? 0 : value;
 
     return uint32(value);
-}
-
-/**
- * @brief Calculates melee critical strike chance gained from agility.
- *
- * @return The melee crit contribution from agility.
- */
-float Player::GetMeleeCritFromAgility()
-{
-    uint32 level = getLevel();
-    uint32 pclass = getClass();
-
-    if (level > GT_MAX_LEVEL) level = GT_MAX_LEVEL;
-
-    GtChanceToMeleeCritBaseEntry const* critBase  = sGtChanceToMeleeCritBaseStore.LookupEntry(pclass - 1);
-    GtChanceToMeleeCritEntry     const* critRatio = sGtChanceToMeleeCritStore.LookupEntry((pclass - 1) * GT_MAX_LEVEL + level - 1);
-    if (critBase == NULL || critRatio == NULL)
-    {
-        return 0.0f;
-    }
-
-    float crit = critBase->base + GetStat(STAT_AGILITY) * critRatio->ratio;
-    return crit * 100.0f;
-}
-
-/**
- * @brief Calculates dodge chance gained from agility.
- *
- * @return The dodge contribution from agility.
- */
-float Player::GetDodgeFromAgility()
-{
-    // Table for base dodge values
-    const float dodge_base[MAX_CLASSES] =
-    {
-        0.0075f,   // Warrior
-        0.00652f,  // Paladin
-        -0.0545f,   // Hunter
-        -0.0059f,   // Rogue
-        0.03183f,  // Priest
-        0.0114f,   // DK
-        0.0167f,   // Shaman
-        0.034575f, // Mage
-        0.02011f,  // Warlock
-        0.0f,      // ??
-        -0.0187f    // Druid
-    };
-    // Crit/agility to dodge/agility coefficient multipliers
-    const float crit_to_dodge[MAX_CLASSES] =
-    {
-        1.1f,      // Warrior
-        1.0f,      // Paladin
-        1.6f,      // Hunter
-        2.0f,      // Rogue
-        1.0f,      // Priest
-        1.0f,      // DK?
-        1.0f,      // Shaman
-        1.0f,      // Mage
-        1.0f,      // Warlock
-        0.0f,      // ??
-        1.7f       // Druid
-    };
-
-    uint32 level = getLevel();
-    uint32 pclass = getClass();
-
-    if (level > GT_MAX_LEVEL) level = GT_MAX_LEVEL;
-
-    // Dodge per agility for most classes equal crit per agility (but for some classes need apply some multiplier)
-    GtChanceToMeleeCritEntry  const* dodgeRatio = sGtChanceToMeleeCritStore.LookupEntry((pclass - 1) * GT_MAX_LEVEL + level - 1);
-    if (dodgeRatio == NULL || pclass > MAX_CLASSES)
-    {
-        return 0.0f;
-    }
-
-    float dodge = dodge_base[pclass-1] + GetStat(STAT_AGILITY) * dodgeRatio->ratio * crit_to_dodge[pclass-1];
-    return dodge * 100.0f;
-}
-
-/**
- * @brief Calculates spell critical strike chance gained from intellect.
- *
- * @return The spell crit contribution from intellect.
- */
-float Player::GetSpellCritFromIntellect()
-{
-    uint32 level = getLevel();
-    uint32 pclass = getClass();
-
-    if (level > GT_MAX_LEVEL) level = GT_MAX_LEVEL;
-
-    GtChanceToSpellCritBaseEntry const* critBase  = sGtChanceToSpellCritBaseStore.LookupEntry(pclass - 1);
-    GtChanceToSpellCritEntry     const* critRatio = sGtChanceToSpellCritStore.LookupEntry((pclass - 1) * GT_MAX_LEVEL + level - 1);
-    if (critBase == NULL || critRatio == NULL)
-    {
-        return 0.0f;
-    }
-
-    float crit = critBase->base + GetStat(STAT_INTELLECT) * critRatio->ratio;
-    return crit * 100.0f;
-}
-
-float Player::GetRatingMultiplier(CombatRating cr) const
-{
-    uint32 level = getLevel();
-
-    if (level > GT_MAX_LEVEL) level = GT_MAX_LEVEL;
-
-    GtCombatRatingsEntry const* Rating = sGtCombatRatingsStore.LookupEntry(cr * GT_MAX_LEVEL + level - 1);
-    if (!Rating)
-    {
-        return 1.0f;                                        // By default use minimum coefficient (not must be called)
-    }
-
-    return 1.0f / Rating->ratio;
-}
-
-float Player::GetRatingBonusValue(CombatRating cr) const
-{
-    return float(GetUInt32Value(PLAYER_FIELD_COMBAT_RATING_1 + cr)) * GetRatingMultiplier(cr);
 }
 
 uint32 Player::GetMeleeCritDamageReduction(uint32 damage) const
@@ -6626,20 +6412,6 @@ uint32 Player::GetDotDamageReduction(uint32 damage) const
     return uint32(spellDot * damage / 100.0f);
 }
 
-float Player::GetExpertiseDodgeOrParryReduction(WeaponAttackType attType) const
-{
-    switch (attType)
-    {
-        case BASE_ATTACK:
-            return GetUInt32Value(PLAYER_EXPERTISE) / 4.0f;
-        case OFF_ATTACK:
-            return GetUInt32Value(PLAYER_OFFHAND_EXPERTISE) / 4.0f;
-        default:
-            break;
-    }
-    return 0.0f;
-}
-
 /**
  * @brief Calculates health regeneration per spirit tick.
  *
@@ -6669,427 +6441,6 @@ float Player::OCTRegenHPPerSpirit()
     float moreSpirit = spirit - baseSpirit;
     float regen = baseSpirit * baseRatio->ratio + moreSpirit * moreRatio->ratio;
     return regen;
-}
-
-/**
- * @brief Calculates mana regeneration per spirit tick.
- *
- * @return The mana regeneration value based on spirit and class.
- */
-float Player::OCTRegenMPPerSpirit()
-{
-    uint32 level = getLevel();
-    uint32 pclass = getClass();
-
-    if (level > GT_MAX_LEVEL) level = GT_MAX_LEVEL;
-
-//    GtOCTRegenMPEntry     const *baseRatio = sGtOCTRegenMPStore.LookupEntry((pclass-1)*GT_MAX_LEVEL + level-1);
-    GtRegenMPPerSptEntry  const* moreRatio = sGtRegenMPPerSptStore.LookupEntry((pclass - 1) * GT_MAX_LEVEL + level - 1);
-    if (moreRatio == NULL)
-    {
-        return 0.0f;
-    }
-
-    // Formula get from PaperDollFrame script
-    float spirit    = GetStat(STAT_SPIRIT);
-    float regen     = spirit * moreRatio->ratio;
-    return regen;
-}
-
-void Player::ApplyRatingMod(CombatRating cr, int32 value, bool apply)
-{
-    m_baseRatingValue[cr] += (apply ? value : -value);
-
-    // explicit affected values
-    switch (cr)
-    {
-        case CR_HASTE_MELEE:
-        {
-            float RatingChange = value * GetRatingMultiplier(cr);
-            ApplyAttackTimePercentMod(BASE_ATTACK, RatingChange, apply);
-            ApplyAttackTimePercentMod(OFF_ATTACK, RatingChange, apply);
-            break;
-        }
-        case CR_HASTE_RANGED:
-        {
-            float RatingChange = value * GetRatingMultiplier(cr);
-            ApplyAttackTimePercentMod(RANGED_ATTACK, RatingChange, apply);
-            break;
-        }
-        case CR_HASTE_SPELL:
-        {
-            float RatingChange = value * GetRatingMultiplier(cr);
-            ApplyCastTimePercentMod(RatingChange, apply);
-            break;
-        }
-        default:
-            break;
-    }
-
-    UpdateRating(cr);
-}
-
-void Player::UpdateRating(CombatRating cr)
-{
-    int32 amount = m_baseRatingValue[cr];
-    if (amount < 0)
-    {
-        amount = 0;
-    }
-    SetUInt32Value(PLAYER_FIELD_COMBAT_RATING_1 + cr, uint32(amount));
-
-    bool affectStats = CanModifyStats();
-
-    switch (cr)
-    {
-        case CR_WEAPON_SKILL:                               // Implemented in Unit::RollMeleeOutcomeAgainst
-        case CR_DEFENSE_SKILL:
-            UpdateDefenseBonusesMod();
-            break;
-        case CR_DODGE:
-            UpdateDodgePercentage();
-            break;
-        case CR_PARRY:
-            UpdateParryPercentage();
-            break;
-        case CR_BLOCK:
-            UpdateBlockPercentage();
-            break;
-        case CR_HIT_MELEE:
-            UpdateMeleeHitChances();
-            break;
-        case CR_HIT_RANGED:
-            UpdateRangedHitChances();
-            break;
-        case CR_HIT_SPELL:
-            UpdateSpellHitChances();
-            break;
-        case CR_CRIT_MELEE:
-            if (affectStats)
-            {
-                UpdateCritPercentage(BASE_ATTACK);
-                UpdateCritPercentage(OFF_ATTACK);
-            }
-            break;
-        case CR_CRIT_RANGED:
-            if (affectStats)
-            {
-                UpdateCritPercentage(RANGED_ATTACK);
-            }
-            break;
-        case CR_CRIT_SPELL:
-            if (affectStats)
-            {
-                UpdateAllSpellCritChances();
-            }
-            break;
-        case CR_HIT_TAKEN_MELEE:                            // Implemented in Unit::MeleeMissChanceCalc
-        case CR_HIT_TAKEN_RANGED:
-            break;
-        case CR_HIT_TAKEN_SPELL:                            // Implemented in Unit::MagicSpellHitResult
-            break;
-        case CR_CRIT_TAKEN_MELEE:                           // Implemented in Unit::RollMeleeOutcomeAgainst (only for chance to crit)
-        case CR_CRIT_TAKEN_RANGED:
-            break;
-        case CR_CRIT_TAKEN_SPELL:                           // Implemented in Unit::SpellCriticalBonus (only for chance to crit)
-            break;
-        case CR_HASTE_MELEE:                                // Implemented in Player::ApplyRatingMod
-        case CR_HASTE_RANGED:
-        case CR_HASTE_SPELL:
-            break;
-        case CR_WEAPON_SKILL_MAINHAND:                      // Implemented in Unit::RollMeleeOutcomeAgainst
-        case CR_WEAPON_SKILL_OFFHAND:
-        case CR_WEAPON_SKILL_RANGED:
-            break;
-        case CR_EXPERTISE:
-            if (affectStats)
-            {
-                UpdateExpertise(BASE_ATTACK);
-                UpdateExpertise(OFF_ATTACK);
-            }
-            break;
-    }
-}
-
-void Player::UpdateAllRatings()
-{
-    for (int cr = 0; cr < MAX_COMBAT_RATING; ++cr)
-    {
-        UpdateRating(CombatRating(cr));
-    }
-}
-
-/**
- * @brief Restores attack timers from currently equipped weapon delays.
- */
-void Player::SetRegularAttackTime()
-{
-    for (int i = 0; i < MAX_ATTACK; ++i)
-    {
-        Item* tmpitem = GetWeaponForAttack(WeaponAttackType(i), true, false);
-        if (tmpitem)
-        {
-            ItemPrototype const* proto = tmpitem->GetProto();
-            if (proto->Delay)
-            {
-                SetAttackTime(WeaponAttackType(i), proto->Delay);
-            }
-            else
-            {
-                SetAttackTime(WeaponAttackType(i), BASE_ATTACK_TIME);
-            }
-        }
-    }
-}
-
-// skill+step, checking for max value
-bool Player::UpdateSkill(uint32 skill_id, uint32 step)
-{
-    if (!skill_id)
-    {
-        return false;
-    }
-
-    SkillStatusMap::iterator itr = mSkillStatus.find(skill_id);
-    if (itr == mSkillStatus.end())
-    {
-        return false;
-    }
-
-    SkillStatusData &skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return false;
-    }
-
-    uint32 valueIndex = PLAYER_SKILL_VALUE_INDEX(skillStatus.pos);
-    uint32 data = GetUInt32Value(valueIndex);
-    uint32 value = SKILL_VALUE(data);
-    uint32 max = SKILL_MAX(data);
-
-    if ((!max) || (!value) || (value >= max))
-    {
-        return false;
-    }
-
-    if (value * 512 < max * urand(0, 512))
-    {
-        uint32 new_value = value + step;
-        if (new_value > max)
-        {
-            new_value = max;
-        }
-
-        SetUInt32Value(valueIndex, MAKE_SKILL_VALUE(new_value, max));
-
-        if (skillStatus.uState != SKILL_NEW)
-        {
-            skillStatus.uState = SKILL_CHANGED;
-        }
-
-        return true;
-    }
-
-    return false;
-}
-
-/**
- * @brief Calculates the configured chance to gain a skill point at the current value.
- *
- * @param SkillValue The player's current skill value.
- * @param GrayLevel The value at which gains become gray.
- * @param GreenLevel The value at which gains become green.
- * @param YellowLevel The value at which gains become yellow.
- * @return The gain chance scaled by ten.
- */
-inline int SkillGainChance(uint32 SkillValue, uint32 GrayLevel, uint32 GreenLevel, uint32 YellowLevel)
-{
-    if (SkillValue >= GrayLevel)
-    {
-        return sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_GREY) * 10;
-    }
-    if (SkillValue >= GreenLevel)
-    {
-        return sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_GREEN) * 10;
-    }
-    if (SkillValue >= YellowLevel)
-    {
-        return sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_YELLOW) * 10;
-    }
-    return sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_ORANGE) * 10;
-}
-
-/**
- * @brief Attempts to increase a crafting skill based on a spell cast.
- *
- * @param spellid The crafting spell identifier.
- * @return True if the skill increased; otherwise, false.
- */
-bool Player::UpdateCraftSkill(uint32 spellid)
-{
-    DEBUG_LOG("UpdateCraftSkill spellid %d", spellid);
-
-    SkillLineAbilityMapBounds bounds = sSpellMgr.GetSkillLineAbilityMapBounds(spellid);
-
-    for (SkillLineAbilityMap::const_iterator _spell_idx = bounds.first; _spell_idx != bounds.second; ++_spell_idx)
-    {
-        SkillLineAbilityEntry const* skill = _spell_idx->second;
-        if (skill->skillId)
-        {
-            uint32 SkillValue = GetPureSkillValue(skill->skillId);
-
-            // Alchemy Discoveries here
-            SpellEntry const* spellEntry = sSpellStore.LookupEntry(spellid);
-            if (spellEntry && spellEntry->Mechanic == MECHANIC_DISCOVERY)
-            {
-                if (uint32 discoveredSpell = GetSkillDiscoverySpell(skill->skillId, spellid, this))
-                {
-                    learnSpell(discoveredSpell, false);
-                }
-            }
-
-            uint32 craft_skill_gain = sWorld.getConfig(CONFIG_UINT32_SKILL_GAIN_CRAFTING);
-
-            return UpdateSkillPro(skill->skillId, SkillGainChance(SkillValue,
-                                  skill->max_value,
-                                  (skill->max_value + skill->min_value) / 2,
-                                  skill->min_value),
-                                  craft_skill_gain);
-        }
-    }
-    return false;
-}
-
-/**
- * @brief Attempts to increase a gathering skill using profession-specific gain rules.
- *
- * @param SkillId The skill identifier to update.
- * @param SkillValue The current skill value.
- * @param RedLevel The red difficulty threshold for the source.
- * @param Multiplicator An additional gain chance multiplier.
- * @return True if the skill increased; otherwise, false.
- */
-bool Player::UpdateGatherSkill(uint32 SkillId, uint32 SkillValue, uint32 RedLevel, uint32 Multiplicator)
-{
-    DEBUG_LOG("UpdateGatherSkill(SkillId %d SkillLevel %d RedLevel %d)", SkillId, SkillValue, RedLevel);
-
-    uint32 gathering_skill_gain = sWorld.getConfig(CONFIG_UINT32_SKILL_GAIN_GATHERING);
-
-    // For skinning and Mining chance decrease with level. 1-74 - no decrease, 75-149 - 2 times, 225-299 - 8 times
-    switch (SkillId)
-    {
-        case SKILL_HERBALISM:
-        case SKILL_LOCKPICKING:
-        case SKILL_JEWELCRAFTING:
-            return UpdateSkillPro(SkillId, SkillGainChance(SkillValue, RedLevel + 100, RedLevel + 50, RedLevel + 25) * Multiplicator, gathering_skill_gain);
-        case SKILL_SKINNING:
-            if (sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_SKINNING_STEPS) == 0)
-            {
-                return UpdateSkillPro(SkillId, SkillGainChance(SkillValue, RedLevel + 100, RedLevel + 50, RedLevel + 25) * Multiplicator, gathering_skill_gain);
-            }
-            else
-            {
-                return UpdateSkillPro(SkillId, (SkillGainChance(SkillValue, RedLevel + 100, RedLevel + 50, RedLevel + 25) * Multiplicator) >> (SkillValue / sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_SKINNING_STEPS)), gathering_skill_gain);
-            }
-        case SKILL_MINING:
-            if (sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_MINING_STEPS) == 0)
-            {
-                return UpdateSkillPro(SkillId, SkillGainChance(SkillValue, RedLevel + 100, RedLevel + 50, RedLevel + 25) * Multiplicator, gathering_skill_gain);
-            }
-            else
-            {
-                return UpdateSkillPro(SkillId, (SkillGainChance(SkillValue, RedLevel + 100, RedLevel + 50, RedLevel + 25) * Multiplicator) >> (SkillValue / sWorld.getConfig(CONFIG_UINT32_SKILL_CHANCE_MINING_STEPS)), gathering_skill_gain);
-            }
-    }
-    return false;
-}
-
-/**
- * @brief Attempts to increase the player's fishing skill.
- *
- * @return True if the skill increased; otherwise, false.
- */
-bool Player::UpdateFishingSkill()
-{
-    DEBUG_LOG("UpdateFishingSkill");
-
-    uint32 SkillValue = GetPureSkillValue(SKILL_FISHING);
-
-    int32 chance = SkillValue < 75 ? 100 : 2500 / (SkillValue - 50);
-
-    uint32 gathering_skill_gain = sWorld.getConfig(CONFIG_UINT32_SKILL_GAIN_GATHERING);
-
-    return UpdateSkillPro(SKILL_FISHING, chance * 10, gathering_skill_gain);
-}
-
-/**
- * @brief Attempts to increase a skill using an explicit percentage chance.
- *
- * @param SkillId The skill identifier to update.
- * @param Chance The gain chance in tenths of a percent.
- * @param step The amount to increase the skill by.
- * @return True if the skill increased; otherwise, false.
- */
-bool Player::UpdateSkillPro(uint16 SkillId, int32 Chance, uint32 step)
-{
-    DEBUG_LOG("UpdateSkillPro(SkillId %d, Chance %3.1f%%)", SkillId, Chance / 10.0);
-    if (!SkillId)
-    {
-        return false;
-    }
-
-    if (Chance <= 0)                                        // speedup in 0 chance case
-    {
-        DEBUG_LOG("Player::UpdateSkillPro Chance=%3.1f%% missed", Chance / 10.0);
-        return false;
-    }
-
-    SkillStatusMap::iterator itr = mSkillStatus.find(SkillId);
-    if (itr == mSkillStatus.end())
-    {
-        return false;
-    }
-
-    SkillStatusData &skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return false;
-    }
-
-    uint32 valueIndex = PLAYER_SKILL_VALUE_INDEX(skillStatus.pos);
-
-    uint32 data = GetUInt32Value(valueIndex);
-    uint16 SkillValue = SKILL_VALUE(data);
-    uint16 MaxValue   = SKILL_MAX(data);
-
-    if (!MaxValue || !SkillValue || SkillValue >= MaxValue)
-    {
-        return false;
-    }
-
-    int32 Roll = irand(1, 1000);
-
-    if (Roll <= Chance)
-    {
-        uint32 new_value = SkillValue + step;
-        if (new_value > MaxValue)
-        {
-            new_value = MaxValue;
-        }
-
-        SetUInt32Value(valueIndex, MAKE_SKILL_VALUE(new_value, MaxValue));
-
-        if (skillStatus.uState != SKILL_NEW)
-        {
-            skillStatus.uState = SKILL_CHANGED;
-        }
-
-        DEBUG_LOG("Player::UpdateSkillPro Chance=%3.1f%% taken", Chance / 10.0);
-        return true;
-    }
-
-    DEBUG_LOG("Player::UpdateSkillPro Chance=%3.1f%% missed", Chance / 10.0);
-    return false;
 }
 
 /**
@@ -7191,477 +6542,6 @@ void Player::UpdateCombatSkills(Unit* pVictim, WeaponAttackType attType, bool de
     {
         return;
     }
-}
-
-/**
- * @brief Modifies the temporary or permanent bonus for a skill.
- *
- * @param skillid The skill identifier to modify.
- * @param val The bonus amount to add or remove.
- * @param talent True for permanent talent-based bonuses; false for temporary bonuses.
- */
-void Player::ModifySkillBonus(uint32 skillid, int32 val, bool talent)
-{
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skillid);
-    if (itr == mSkillStatus.end() || itr->second.uState == SKILL_DELETED)
-    {
-        return;
-    }
-
-    uint32 bonusIndex = PLAYER_SKILL_BONUS_INDEX(itr->second.pos);
-
-    uint32 bonus_val = GetUInt32Value(bonusIndex);
-    int16 temp_bonus = SKILL_TEMP_BONUS(bonus_val);
-    int16 perm_bonus = SKILL_PERM_BONUS(bonus_val);
-
-    if (talent)                                         // permanent bonus stored in high part
-    {
-        SetUInt32Value(bonusIndex, MAKE_SKILL_BONUS(temp_bonus, perm_bonus + val));
-    }
-    else                                                // temporary/item bonus stored in low part
-    {
-        SetUInt32Value(bonusIndex, MAKE_SKILL_BONUS(temp_bonus + val, perm_bonus));
-    }
-}
-
-/**
- * @brief Updates level-scaled skills to match the player's current level cap.
- */
-void Player::UpdateSkillsForLevel()
-{
-    uint16 maxconfskill = sWorld.GetConfigMaxSkillValue();
-    uint32 maxSkill = GetMaxSkillValueForLevel();
-
-    bool alwaysMaxSkill = sWorld.getConfig(CONFIG_BOOL_ALWAYS_MAX_SKILL_FOR_LEVEL);
-
-    for (SkillStatusMap::iterator itr = mSkillStatus.begin(); itr != mSkillStatus.end(); ++itr)
-    {
-        SkillStatusData &skillStatus = itr->second;
-        if (skillStatus.uState == SKILL_DELETED)
-        {
-            continue;
-        }
-
-        uint32 pskill = itr->first;
-
-        SkillLineEntry const* pSkill = sSkillLineStore.LookupEntry(pskill);
-        if (!pSkill)
-        {
-            continue;
-        }
-
-        if (GetSkillRangeType(pSkill, false) != SKILL_RANGE_LEVEL)
-        {
-            continue;
-        }
-
-        uint32 valueIndex = PLAYER_SKILL_VALUE_INDEX(skillStatus.pos);
-        uint32 data = GetUInt32Value(valueIndex);
-        uint32 max = SKILL_MAX(data);
-        uint32 val = SKILL_VALUE(data);
-
-        /// update only level dependent max skill values
-        if (max != 1)
-        {
-            /// maximize skill always
-            if (alwaysMaxSkill)
-            {
-                SetUInt32Value(valueIndex, MAKE_SKILL_VALUE(maxSkill, maxSkill));
-                if (skillStatus.uState != SKILL_NEW)
-                {
-                    skillStatus.uState = SKILL_CHANGED;
-                }
-            }
-            else if (max != maxconfskill)                   /// update max skill value if current max skill not maximized
-            {
-                SetUInt32Value(valueIndex, MAKE_SKILL_VALUE(val, maxSkill));
-                if (skillStatus.uState != SKILL_NEW)
-                {
-                    skillStatus.uState = SKILL_CHANGED;
-                }
-            }
-        }
-    }
-}
-
-/**
- * @brief Raises non-profession skills to their current maximum values.
- */
-void Player::UpdateSkillsToMaxSkillsForLevel()
-{
-    for (SkillStatusMap::iterator itr = mSkillStatus.begin(); itr != mSkillStatus.end(); ++itr)
-    {
-        SkillStatusData &skillStatus = itr->second;
-        if (skillStatus.uState == SKILL_DELETED)
-        {
-            continue;
-        }
-
-        uint32 pskill = itr->first;
-        if (IsProfessionOrRidingSkill(pskill))
-        {
-            continue;
-        }
-        uint32 valueIndex = PLAYER_SKILL_VALUE_INDEX(skillStatus.pos);
-        uint32 data = GetUInt32Value(valueIndex);
-
-        uint32 max = SKILL_MAX(data);
-
-        if (max > 1)
-        {
-            SetUInt32Value(valueIndex, MAKE_SKILL_VALUE(max, max));
-            if (skillStatus.uState != SKILL_NEW)
-            {
-                skillStatus.uState = SKILL_CHANGED;
-            }
-        }
-
-        if (pskill == SKILL_DEFENSE)
-        {
-            UpdateDefenseBonusesMod();
-        }
-    }
-}
-
-// This functions sets a skill line value (and adds if doesn't exist yet)
-// To "remove" a skill line, set it's values to zero
-void Player::SetSkill(uint16 id, uint16 currVal, uint16 maxVal, uint16 step /*=0*/)
-{
-    if (!id)
-    {
-        return;
-    }
-
-    SkillStatusMap::iterator itr = mSkillStatus.find(id);
-
-    // has skill
-    if (itr != mSkillStatus.end() && itr->second.uState != SKILL_DELETED)
-    {
-        SkillStatusData &skillStatus = itr->second;
-        if (currVal)
-        {
-            if (step)                                      // need update step
-            {
-                SetUInt32Value(PLAYER_SKILL_INDEX(skillStatus.pos), MAKE_PAIR32(id, step));
-            }
-
-            // update value
-            SetUInt32Value(PLAYER_SKILL_VALUE_INDEX(skillStatus.pos), MAKE_SKILL_VALUE(currVal, maxVal));
-            if (skillStatus.uState != SKILL_NEW)
-            {
-                skillStatus.uState = SKILL_CHANGED;
-            }
-            // learnSkillRewardedSpells(id, currVal);       // pre-3.x have only 1 skill level req (so at learning only)
-        }
-        else                                                // remove
-        {
-            // clear skill fields
-            SetUInt32Value(PLAYER_SKILL_INDEX(skillStatus.pos), 0);
-            SetUInt32Value(PLAYER_SKILL_VALUE_INDEX(skillStatus.pos), 0);
-            SetUInt32Value(PLAYER_SKILL_BONUS_INDEX(skillStatus.pos), 0);
-
-            // mark as deleted or simply remove from map if not saved yet
-            if (skillStatus.uState != SKILL_NEW)
-            {
-                skillStatus.uState = SKILL_DELETED;
-            }
-            else
-            {
-                mSkillStatus.erase(itr);
-            }
-
-            // remove all spells that related to this skill
-            for (uint32 j = 0; j < sSkillLineAbilityStore.GetNumRows(); ++j)
-            {
-                if (SkillLineAbilityEntry const* pAbility = sSkillLineAbilityStore.LookupEntry(j))
-                    if (pAbility->skillId == id)
-                    {
-                        removeSpell(sSpellMgr.GetFirstSpellInChain(pAbility->spellId));
-                    }
-            }
-        }
-    }
-    else if (currVal)                                       // add
-    {
-        for (int i = 0; i < PLAYER_MAX_SKILLS; ++i)
-        {
-            if (!GetUInt32Value(PLAYER_SKILL_INDEX(i)))
-            {
-                SkillLineEntry const* pSkill = sSkillLineStore.LookupEntry(id);
-                if (!pSkill)
-                {
-                    sLog.outError("Skill not found in SkillLineStore: skill #%u", id);
-                    return;
-                }
-
-                SetUInt32Value(PLAYER_SKILL_INDEX(i), MAKE_PAIR32(id, step));
-                SetUInt32Value(PLAYER_SKILL_VALUE_INDEX(i), MAKE_SKILL_VALUE(currVal, maxVal));
-
-                // insert new entry or update if not deleted old entry yet
-                if (itr != mSkillStatus.end())
-                {
-                    itr->second.pos = i;
-                    itr->second.uState = SKILL_CHANGED;
-                }
-                else
-                {
-                    mSkillStatus.insert(SkillStatusMap::value_type(id, SkillStatusData(i, SKILL_NEW)));
-                }
-
-                // apply skill bonuses
-                SetUInt32Value(PLAYER_SKILL_BONUS_INDEX(i), 0);
-
-                // temporary bonuses
-                AuraList const& mModSkill = GetAurasByType(SPELL_AURA_MOD_SKILL);
-                for (AuraList::const_iterator j = mModSkill.begin(); j != mModSkill.end(); ++j)
-                {
-                    if ((*j)->GetModifier()->m_miscvalue == int32(id))
-                    {
-                        (*j)->ApplyModifier(true);
-                    }
-                }
-
-                // permanent bonuses
-                AuraList const& mModSkillTalent = GetAurasByType(SPELL_AURA_MOD_SKILL_TALENT);
-                for (AuraList::const_iterator j = mModSkillTalent.begin(); j != mModSkillTalent.end(); ++j)
-                {
-                    if ((*j)->GetModifier()->m_miscvalue == int32(id))
-                    {
-                        (*j)->ApplyModifier(true);
-                    }
-                }
-
-                // Learn all spells for skill
-                learnSkillRewardedSpells(id, currVal);
-                return;
-            }
-        }
-    }
-}
-
-/**
- * @brief Checks whether the player currently has a skill line.
- *
- * @param skill The skill identifier to test.
- * @return True if the skill exists and is not deleted; otherwise, false.
- */
-bool Player::HasSkill(uint32 skill) const
-{
-    if (!skill)
-    {
-        return false;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    return (itr != mSkillStatus.end() && itr->second.uState != SKILL_DELETED);
-}
-
-/**
- * @brief Gets the total current value of a skill including bonuses.
- *
- * @param skill The skill identifier to query.
- * @return The current effective skill value.
- */
-uint16 Player::GetSkillValue(uint32 skill) const
-{
-    if (!skill)
-    {
-        return 0;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    if (itr == mSkillStatus.end())
-    {
-        return 0;
-    }
-
-    SkillStatusData const& skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return 0;
-    }
-
-    uint32 bonus = GetUInt32Value(PLAYER_SKILL_BONUS_INDEX(skillStatus.pos));
-
-    int32 result = int32(SKILL_VALUE(GetUInt32Value(PLAYER_SKILL_VALUE_INDEX(skillStatus.pos))));
-    result += SKILL_TEMP_BONUS(bonus);
-    result += SKILL_PERM_BONUS(bonus);
-    return result < 0 ? 0 : result;
-}
-
-/**
- * @brief Gets the total maximum value of a skill including bonuses.
- *
- * @param skill The skill identifier to query.
- * @return The effective maximum skill value.
- */
-uint16 Player::GetMaxSkillValue(uint32 skill) const
-{
-    if (!skill)
-    {
-        return 0;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    if (itr == mSkillStatus.end())
-    {
-        return 0;
-    }
-
-    SkillStatusData const& skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return 0;
-    }
-
-    uint32 bonus = GetUInt32Value(PLAYER_SKILL_BONUS_INDEX(skillStatus.pos));
-
-    int32 result = int32(SKILL_MAX(GetUInt32Value(PLAYER_SKILL_VALUE_INDEX(skillStatus.pos))));
-    result += SKILL_TEMP_BONUS(bonus);
-    result += SKILL_PERM_BONUS(bonus);
-    return result < 0 ? 0 : result;
-}
-
-/**
- * @brief Gets the unmodified maximum value of a skill.
- *
- * @param skill The skill identifier to query.
- * @return The stored maximum skill value without bonuses.
- */
-uint16 Player::GetPureMaxSkillValue(uint32 skill) const
-{
-    if (!skill)
-    {
-        return 0;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    if (itr == mSkillStatus.end())
-    {
-        return 0;
-    }
-
-    SkillStatusData const& skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return 0;
-    }
-
-    return SKILL_MAX(GetUInt32Value(PLAYER_SKILL_VALUE_INDEX(skillStatus.pos)));
-}
-
-/**
- * @brief Gets the base value of a skill including permanent bonuses.
- *
- * @param skill The skill identifier to query.
- * @return The base skill value with permanent bonuses applied.
- */
-uint16 Player::GetBaseSkillValue(uint32 skill) const
-{
-    if (!skill)
-    {
-        return 0;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    if (itr == mSkillStatus.end())
-    {
-        return 0;
-    }
-
-    SkillStatusData const& skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return 0;
-    }
-
-    int32 result = int32(SKILL_VALUE(GetUInt32Value(PLAYER_SKILL_VALUE_INDEX(skillStatus.pos))));
-    result += SKILL_PERM_BONUS(GetUInt32Value(PLAYER_SKILL_BONUS_INDEX(skillStatus.pos)));
-    return result < 0 ? 0 : result;
-}
-
-/**
- * @brief Gets the raw stored value of a skill without bonuses.
- *
- * @param skill The skill identifier to query.
- * @return The raw stored skill value.
- */
-uint16 Player::GetPureSkillValue(uint32 skill) const
-{
-    if (!skill)
-    {
-        return 0;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    if (itr == mSkillStatus.end())
-    {
-        return 0;
-    }
-
-    SkillStatusData const& skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return 0;
-    }
-
-    return SKILL_VALUE(GetUInt32Value(PLAYER_SKILL_VALUE_INDEX(skillStatus.pos)));
-}
-
-/**
- * @brief Gets the permanent bonus value applied to a skill.
- *
- * @param skill The skill identifier to query.
- * @return The permanent skill bonus.
- */
-int16 Player::GetSkillPermBonusValue(uint32 skill) const
-{
-    if (!skill)
-    {
-        return 0;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    if (itr == mSkillStatus.end())
-    {
-        return 0;
-    }
-
-    SkillStatusData const& skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return 0;
-    }
-
-    return SKILL_PERM_BONUS(GetUInt32Value(PLAYER_SKILL_BONUS_INDEX(skillStatus.pos)));
-}
-
-/**
- * @brief Gets the temporary bonus value applied to a skill.
- *
- * @param skill The skill identifier to query.
- * @return The temporary skill bonus.
- */
-int16 Player::GetSkillTempBonusValue(uint32 skill) const
-{
-    if (!skill)
-    {
-        return 0;
-    }
-
-    SkillStatusMap::const_iterator itr = mSkillStatus.find(skill);
-    if (itr == mSkillStatus.end())
-    {
-        return 0;
-    }
-
-    SkillStatusData const& skillStatus = itr->second;
-    if (skillStatus.uState == SKILL_DELETED)
-    {
-        return 0;
-    }
-
-    return SKILL_TEMP_BONUS(GetUInt32Value(PLAYER_SKILL_BONUS_INDEX(skillStatus.pos)));
 }
 
 /* Called from Player::SendInitialPacketsBeforeAddToMap */
